feat: integrity linter minimal working version
Check for keep track of calls to Marshal() and Unmarshal() on integrity
option REQUIRED  proto messages.
Check that Sign is called before Marshals() on instances of these messages.
Check that Verify is called after Unmarshals() on insances of these messages.
Report diagnostic on Marshalls() and Unmarshals() that don't satisfy this.
Add Extra examples to tests to show functionality.