feat: add handing pem keypairs and accessing them via keyID
Private public keypairs have a keyID which is used to identify them
so that the correct public key can be used to verify the signature
signed by the private key.
Add keypairs.go which stores these util functions for loading from
files and for moving the test-keys from their dir to the dir used
message-integrity.